Hotfix auto logout feature
