+++ src/components/EditForm/EditForm.tsx
@@ -63,10 +63,6 @@ export const EditFormComponent: React.FC<EditFormProps> = ({
   }, []);
 
   const onSubmit: SubmitHandler<InputListType> = async data => {
-    axiosClient.get(workerUrl.getMeWorker()).catch(function () {
-      toast.error('로그인 정보가 일치하지 않아요\n자동으로 로그아웃 됩니다');
-      handleLogout();
-    });
     const userReqData: UserEditReqData = {
       name: data.name,
       email: data.email,
@@ -128,9 +124,15 @@ export const EditFormComponent: React.FC<EditFormProps> = ({
       .then(function () {
         setCompanyFormModalVisible(true);
       })
-      .catch(function () {
-        toast.error('로그인 정보가 일치하지 않아요\n자동으로 로그아웃 됩니다');
-        handleLogout();
+      .catch(function (error: AxiosError) {
+        if (error?.response?.status === 401) {
+          toast.error(
+            '로그인 정보가 일치하지 않아요\n자동으로 로그아웃 됩니다',
+          );
+          handleLogout();
+        } else {
+          setCompanyFormModalVisible(true);
+        }
       });
   };
