+++ src/components/SideBar/SideBar.tsx
@@ -10,16 +10,14 @@ import { Form } from 'components/Form';
 import { VerifyForm } from 'components/VerifyForm';
 import { SideBarButton } from 'components/common/SideBarButton';
 import { EditForm } from 'components/EditForm';
-import { Burger } from 'assets/icons/Burger';
-import { Cancel } from 'assets/icons/Cancel';
-import { Logo } from 'assets/icons/Logo';
 import axiosClient from 'libs/axios/axiosClient';
 import { workerUrl } from 'libs/api/apiUrlControllers';
 import { UserRule } from 'types/site.type';
 
 import * as S from './SideBar.styles';
-import { handleAuth, handleLogout } from './container';
+import { handleLogout } from './container';
 import { Header } from 'components/common/Header';
+import { AxiosError } from 'axios';
 
 interface SideBarProps {
   cookies: {
@@ -80,9 +78,15 @@ export const SideBarComponent: React.FC<SideBarProps> = ({ cookies }) => {
       .then(function () {
         setModalVisible(true);
       })
-      .catch(function () {
-        toast.error('로그인 정보가 일치하지 않아요\n자동으로 로그아웃 됩니다');
-        handleLogout();
+      .catch(function (error: AxiosError) {
+        if (error?.response?.status === 401) {
+          toast.error(
+            '로그인 정보가 일치하지 않아요\n자동으로 로그아웃 됩니다',
+          );
+          handleLogout();
+        } else {
+          setModalVisible(true);
+        }
       });
   };
 
@@ -92,39 +96,21 @@ export const SideBarComponent: React.FC<SideBarProps> = ({ cookies }) => {
       .then(function () {
         setVerifyFormModalVisible(true);
       })
-      .catch(function () {
-        toast.error('로그인 정보가 일치하지 않아요\n자동으로 로그아웃 됩니다');
-        handleLogout();
-      });
-  };
-
-  const handleMenuClick = () => {
-    axiosClient
-      .get(workerUrl.getMeWorker())
-      .then(function () {
-        setMenuVisible(true);
-      })
-      .catch(function () {
-        toast.error('로그인 정보가 일치하지 않아요\n자동으로 로그아웃 됩니다');
-        handleLogout();
+      .catch(function (error: AxiosError) {
+        if (error?.response?.status === 401) {
+          toast.error(
+            '로그인 정보가 일치하지 않아요\n자동으로 로그아웃 됩니다',
+          );
+          handleLogout();
+        } else {
+          setVerifyFormModalVisible(true);
+        }
       });
   };
 
   return (
     <>
       <S.SideBar>
-        <S.SideBarHeader>
-          <Logo logoColor="blue" />
-          {userRules === 'NO_AUTH_USER' ? (
-            <S.SignUpAnchor onClick={() => handleAuth()}>
-              회원가입/로그인
-            </S.SignUpAnchor>
-          ) : (
-            <div onClick={() => handleMenuClick()}>
-              {menuVisible ? <Cancel /> : <Burger />}
-            </div>
-          )}
-        </S.SideBarHeader>
         <Header
           userRules={userRules}
           setMenuVisible={setMenuVisible}
