+++ src/components/common/Header/Header.tsx
@@ -1,12 +1,16 @@
 import { Dispatch, SetStateAction } from 'react';
+import toast from 'react-hot-toast';
 
 import { Logo } from 'assets/icons/Logo';
 import { Cancel } from 'assets/icons/Cancel';
 import { Burger } from 'assets/icons/Burger';
+import axiosClient from 'libs/axios/axiosClient';
+import { workerUrl } from 'libs/api/apiUrlControllers';
 import { UserRule } from 'types/site.type';
 
 import * as S from './Header.styles';
-import { handleAuth } from './container';
+import { handleAuth, handleLogout } from './container';
+import { AxiosError } from 'axios';
 
 interface HeaderProps {
   userRules: UserRule;
@@ -19,6 +23,24 @@ export const HeaderComponent: React.FC<HeaderProps> = ({
   menuVisible,
   setMenuVisible,
 }) => {
+  const handleMenuClick = () => {
+    axiosClient
+      .get(workerUrl.getMeWorker())
+      .then(function () {
+        setMenuVisible(true);
+      })
+      .catch(function (error: AxiosError) {
+        if (error?.response?.status === 401) {
+          toast.error(
+            '로그인 정보가 일치하지 않아요\n자동으로 로그아웃 됩니다',
+          );
+          handleLogout();
+        } else {
+          setMenuVisible(true);
+        }
+      });
+  };
+
   return (
     <S.Header>
       <Logo logoColor="blue" />
@@ -27,11 +49,7 @@ export const HeaderComponent: React.FC<HeaderProps> = ({
           회원가입/로그인
         </S.HeaderAnchor>
       ) : (
-        <div
-          onClick={() => {
-            setMenuVisible(prev => !prev);
-          }}
-        >
+        <div onClick={() => handleMenuClick()}>
           {menuVisible ? <Cancel /> : <Burger />}
         </div>
       )}
